feat: total nmb repos and nmb dirty repos
Added two new attributes to instances of Repositories. One for the total
number of found repositories and one for the number of dirty
repositories.

With it, there is a new .summary() method printing a quick report. This
method is currently invoked at the very end of __main__.py and the
application script.