fix: tests pass now also on GitHub actions
Reason for not passing: tested code attempts to read a file which is not
present, hence it falls back to defaults as intended. However, the test
cases in question test successful config reads (file present, no syntax
error in config). As the configuration was present on the local test
system, the tests passed. However, on GitHub actions that file is not
present. To remedy this, open() is now mocked, hence the code progresses
as expected now during testing. The open() operation is not further
scrutinized as its not the intention to test whether builtin tools
function properly. The same goes for tomllib.load. Hence, only the
end result, namely of tomllib.load() is patched.